Move ReferenceIndex into a separate module
